Variables are now uint16_t (previous uint8_t)

+++ firmware/src/adc.h
@@ -24,7 +24,7 @@ typedef enum adc_channels{
 } adc_channels_t;                           //*< the adc_channel type
 extern volatile uint8_t print_adc;
 
-extern volatile uint8_t batvoltage, position, batcurrent;
+extern volatile uint16_t batvoltage, position, batcurrent;
 
 void adc_init(void);
